refactor(wiki): split 1454-line wiki-generator into pure builders
The wiki generator mixed 30+ pure builder functions (each turning
domain data into a markdown body) with disk I/O, manifest diffing,
and the orchestrator pipeline. Split into focused modules under
core/services/wiki/ and reduced wiki-generator.ts from 1454 to
298 lines (just the orchestrator + Obsidian seeding).

Modules created:
  - wiki/_shared.ts              — slugify/sha256/chunkEntries + types
  - wiki/memory-builder.ts       — memory/<type>.md + tags/<key>.md + ships
  - wiki/llm-analysis-builder.ts — patterns/architecture/tech-debt/insights
  - wiki/concept-builder.ts      — analysis/<kind>/<slug>.md + history.md
  - wiki/release-builder.ts      — releases/* (parses CHANGELOG.md)
  - wiki/workflow-builder.ts     — workflows/<command>.md
  - wiki/index-builder.ts        — top-level index.md
  - wiki/manifest.ts             — readManifest, write/remove/sweepStale
  - wiki/fingerprint.ts          — input-state hash for short-circuit

All builders are pure: take domain values in, return Map<rel, body>
out. The orchestrator owns I/O — Promise.all over storage queries,
manifest diff loop, atomic fingerprint stamp last.

Behavior unchanged:
  - typecheck clean
  - biome clean
  - 1029 tests pass (3477 expects)
  - bun run build succeeds
